Finished Travel Journey App - moved it to different folder
