+++ index.html
@@ -4,13 +4,13 @@
 	<link rel="preconnect" href="https://fonts.googleapis.com">
 	<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 	<link href="https://fonts.googleapis.com/css2?family=Inter:wght@100..900&display=swap" rel="stylesheet">
-	<link rel="stylesheet" href="./src/index.css">
+	<link rel="stylesheet" href="/src/travel-journal-project/TravelIndex.css">
 </head>
 
 <body>
 	<div id="root"></div>
 	<!-- <script src="./src/for-understanding/f-index.jsx" type="module"></script> -->
-	<script src="./src/index.jsx" type="module"></script>
+	<script src="/src/travel-journal-project/TravelIndex.jsx" type="module"></script>
 </body>
 
 </html>
