+++ src/travel-journal-project/TravelIndex.css
@@ -1,0 +1,91 @@
+html, body {
+  margin: 0;
+  font-family: 'Inter';
+}
+
+#root {
+ align-items: center;
+ justify-items: center;
+ overflow-x: hidden;
+}
+
+.container {
+  margin: 45px 40px;
+  display: flex;
+  flex-direction: column;
+  gap: 40px;
+}
+
+header {
+  background: #F55A5A;
+  display:flex; 
+  width: 100%;
+  align-items: center;
+  justify-content: center;
+  color: white;
+  height: 55px;
+  gap: 8px;
+}
+
+header > h1 {
+  font-size: 1rem;
+  font-weight:500;
+  font-family: 'Inter';
+}
+
+article.journal-entry {
+  display: flex;
+}
+
+article.journal-entry .marker {
+  width: 7px;
+  margin-right: 4px;
+}
+
+article.journal-entry .main-image-container {
+  width: 125px;
+  height: 168px;
+  overflow: hidden;
+  border-radius: 5px;
+  flex-shrink: 0;
+  margin-right: 19px;
+}
+
+article.journal-entry .main-image {
+  height: 100%;
+  width: 100%;
+  object-fit: cover;
+}
+
+.info-container span.country {
+  text-transform: uppercase;
+  font-size: 0.7rem;
+}
+
+.info-container a {
+  color: #2B283A;
+  font-size: 0.7rem;
+}
+
+.info-container a:visited {
+  color: #918E9B;
+}
+
+.info-container p.trip-dates {
+  font-weight: bold;
+}
+
+.info-container p.entry-text {
+  line-height: 1.5;
+  font-size: 0.8rem;
+  display: -webkit-box;
+  -webkit-line-clamp: 3;
+  -webkit-box-orient: vertical;  
+  overflow: hidden;
+}
+
+.info-country {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
