+++ src/travel-journal-project/TravelIndex.jsx
@@ -1,0 +1,7 @@
+import {createRoot} from "react-dom/client"
+import App  from "./TravelApp"
+
+const root = createRoot(document.getElementById("root"))
+root.render(
+  <App/>
+)
